requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
Click==7.0
contourpy==1.1.0
crispy-bootstrap4==2022.1
cryptography==41.0.2
cycler==0.11.0
Django==4.1
django-bootstrap4==23.1
django-crispy-forms==2.0
django-daraja==1.3.0
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
fonttools==4.41.0
graphviz==0.20.1
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
phonenumbers==8.13.16
Pillow==10.0.0
ply==3.11
progressbar2==4.2.0
psycopg2-binary==2.9.6
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-utils==3.7.0
pytz==2023.3
PyYAML==6.0.1
quantiphy==2.19
reportlab==4.0.4
requests==2.31.0
scipy==1.11.1
setuptools-scm==7.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tabulate==0.9.0
typing_extensions==4.6.3
urllib3==2.0.4
zero==0.9.1
